update(get, post, put and delete routes)

+++ controllers/books.js
@@ -1,4 +1,5 @@
 const Book = require('../models/Book')
+const fs = require('fs')
 
 exports.getAllBooks = (req, res, next) => {
     Book.find()
@@ -13,30 +14,73 @@ exports.getOneBook = (req, res, next) => {
 }
 
 exports.createBook = (req, res, next) => {
+    const bookObject = JSON.parse(req.body.book)
+    delete bookObject._id
+    delete bookObject.userId
     const book = new Book({
-        title: req.body.title,
-        description: req.body.description,
-        imageUrl: req.body.imageUrl,
-        price: req.body.price,
-        userId: req.body.userId,
+        ...bookObject,
+        userId: req.auth.userId,
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`,
     })
     book.save()
         .then(() => {
-            res.status(201).json({ message: 'Book saved successfully!' })
+            res.status(201).json({ message: 'Livre enregistré' })
         })
         .catch((error) => {
-            res.status(400).json({ error: error })
+            res.status(400).json({ error })
         })
 }
 
+exports.bestRatingBooks = (req, res, next) => {}
+
 exports.modifyBook = (req, res, next) => {
-    Book.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
-        .then(() => res.status(200).json({ message: 'Objet modifié !' }))
-        .catch((error) => res.status(400).json({ error }))
+    const bookObject = req.file
+        ? {
+              ...JSON.parse(req.body.book),
+              imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`,
+          }
+        : { ...req.body }
+
+    delete bookObject._userId
+    Book.findOne({ _id: req.params.id })
+        .then((book) => {
+            if (book.userId != req.auth.userId) {
+                res.status(401).json({ message: 'Not authorized' })
+            } else {
+                Book.updateOne(
+                    { _id: req.params.id },
+                    { ...bookObject, _id: req.params.id }
+                )
+                    .then(() =>
+                        res.status(200).json({ message: 'Objet modifié!' })
+                    )
+                    .catch((error) => res.status(401).json({ error }))
+            }
+        })
+        .catch((error) => {
+            res.status(400).json({ error })
+        })
 }
 
 exports.deleteBook = (req, res, next) => {
-    Book.deleteOne({ _id: req.params.id })
-        .then(() => res.status(200).json({ message: 'Objet supprimé !' }))
-        .catch((error) => res.status(400).json({ error }))
+    Book.findOne({ _id: req.params.id })
+        .then((book) => {
+            if (book.userId != req.auth.userId) {
+                res.status(401).json({ message: 'Not authorized' })
+            } else {
+                const filename = book.imageUrl.split('/images/')[1]
+                fs.unlink(`images/${filename}`, () => {
+                    Book.deleteOne({ _id: req.params.id })
+                        .then(() => {
+                            res.status(200).json({
+                                message: 'Objet supprimé !',
+                            })
+                        })
+                        .catch((error) => res.status(401).json({ error }))
+                })
+            }
+        })
+        .catch((error) => {
+            res.status(500).json({ error })
+        })
 }
